requirements.txt
openai>=1.0.0
python-docx==1.0.1
SpeechRecognition==3.10.0
pydub==0.25.1
python-dotenv==1.0.0